wip: rely on git version of prettyplease, refactor autofmt
