+++ packages/autofmt/src/children.rs
@@ -1,0 +1,4 @@
+// write the body of a component or element
+pub fn write_children() {
+
+}
