--- packages/autofmt/src/prettyplease/lifetime.rs
@@ -1,9 +1,0 @@
-use super::algorithm::Printer;
-use syn::Lifetime;
-
-impl Printer {
-    pub fn lifetime(&mut self, lifetime: &Lifetime) {
-        self.word("'");
-        self.ident(&lifetime.ident);
-    }
-}
